Refactor API-related components and introduce `ApiBundle`
This commit moves API components, such as adapter, attributes and exception classes, from the `App` namespace to a new `Module\Api` namespace. An `ApiBundle` is created to manage the services, responses and event listeners related to API functionality. This restructuring improves organization and makes the API-related code more modular and easier to maintain.