Refactor ApiResponse and update handling of HTTP status codes
This commit includes the refactoring of ApiResponse and changes the handling of HTTP status codes. ApiResponse is moved to the DTO directory. In the process, checks for invalid status codes when data is null are removed and HttpStatus enum is introduced for managing HTTP status codes in a more structured manner. Status code handling in ApiResponse is updated to use the new HttpStatus enum. Changes in various classes to use HttpStatus and several improvements in doc comments and annotations have also been included.