Refactor code and add phparkitect for architecture rules enforcement
This commit introduces phparkitect, a tool that enforces architecture rules, and it also includes significant code refactoring. The modifications mostly involve renaming various 'Enum' and 'Trait' files to match uniform naming conventions as part of implementing the phparkitect rules. Another noticeable change is the update in namespaces and class references in multiple files to reflect the new file names. Furthermore, a new configuration file for phparkitect was added to define the architecture rules thus ensuring that the project's structure and design are easily maintainable.