Reorganize documentation and update CI pipeline
In this commit, the home page of the documentation was moved from 'home.md' to 'index.md', improving consistency. A new stage 'deploy' was added to the .gitlab-ci.yml file to build mkdocs with Python, enhancing the deployment process. An update was also made to 'mkdocs.yml' to reflect the name changes.