Add Sentry package and enable error tracking
The Sentry package has been added to the bundles and composer.json. This allows for improved error tracking in both development (`dev`) and production (`prod`). Additionally, required settings for Sentry have been added in a new sentry.yaml file and in the .env file. Changes also include the addition of `nyholm/psr7` to composer.json and an adjustment to the .docker/php/Dockerfile.